Fixing the Mess:
This commit did quite a few. Went through and fixed some of the names to match a reasonable naming convention. Also updated RobotBuilder yml file which was, for some reason, not merging properly. It is now set up so that you can export the code from RobotBuilder and build the project without having to change any of this code.

This has been tested and the following works:
Intake - Roller & Solenoid turn on and off by switch
         Roller outtakes when held
Chassis - Drives six motors in tank drive.
Shifter - Shifts low to high gear.
Compressor - Compresses to an automatic shutoff at 120 psi.