RoboRealm Vision Processing 4
Stuff works. We display the distance on the SmartDashboard. The variables Distance, BLOB_COUNT, and HotStatus are sent via Network Tables. BLOB_COUNT and HotStatus basically do the same thing, but HotStatus may be more convenient. If it is = to 0, no tape/blobs are seen, if it is = 1 one piece of tape is seen. If HotStatus is = 2 that means the goal is hot as it senses two blobs, which have a very high probability of being both pieces of tape.