RoboRealm Vision Processing 2/14/14
Small fix, HotStatus was not being sent to the Smart Dashboard.

TO RUN THIS!:
Make sure you have RoboRealm installed. Open the file, make sure the Axis_Internet_Camera 10.7.1.11 is not disabled, Blob_Seperate is an example of a disabled object. Connect to the robot and be sure the Run button is pushed in on RoboRealm at the top. Seek help if the Axis Camera doesn't send images.